requirements.txt
streamlit
langchain
openai
chromadb
python-dotenv
sentence-transformers
PyMuPDF
Pillow